src/backend/RecommendationService/requirements.txt
flask
surprise
pandas
psycopg2-binary
python-dotenv